Properly fix VS2019 compile error.

+++ tools/ktx/command.cpp
@@ -55,7 +55,8 @@ bool Command::launchDebugger()
 {
     // Get System directory, typically c:\windows\system32
     std::wstring systemDir(MAX_PATH + 1, '\0');
-    UINT nChars = GetSystemDirectoryW(&systemDir[0], systemDir.length());
+    UINT nChars = GetSystemDirectoryW(&systemDir[0],
+                                static_cast<UINT>(systemDir.length()));
     if (nChars == 0) return false; // failed to get system directory
     systemDir.resize(nChars);
